Remove unused vinxi dependency and convert remaining JS/JSX to TypeScript
Vinxi was left in package.json after the TanStack Start migration but is not
imported or invoked anywhere. Drops it from dependencies and the lockfile.

Also converts the remaining JavaScript/JSX sources to TypeScript: all
components under src/, portfolio data, the components barrel, and the
vitest/playwright/tailwind configs. server.js and postcss.config.cjs stay as
CommonJS/Node entry points.
