+++ package.json
@@ -21,8 +21,7 @@
     "react": "^19.2.5",
     "react-dom": "^19.2.5",
     "react-icons": "^5.6.0",
-    "react-p5": "^1.4.1",
-    "vinxi": "^0.5.1"
+    "react-p5": "^1.4.1"
   },
   "devDependencies": {
     "@tanstack/router-plugin": "^1.95.0",
