+++ src/components/Navbar/Navbar.tsx
@@ -1,10 +1,15 @@
-import React from "react";
 import "./navbar.css";
 import { AnimatePresence, motion } from "framer-motion";
 import { Link } from "@tanstack/react-router";
 import { FaBars } from "react-icons/fa6";
 
-function Navbar({ hideLogo, toggleDrawer, isScrolledDown }) {
+type NavbarProps = {
+  hideLogo?: boolean;
+  toggleDrawer: () => void;
+  isScrolledDown: boolean;
+};
+
+function Navbar({ hideLogo, toggleDrawer, isScrolledDown }: NavbarProps) {
   return (
     <div className={`navbar${isScrolledDown ? " navbar--scrolled" : ""}`}>
       <div className="navbar__left">
