--- src/components/index.js
@@ -1,3 +1,0 @@
-export {default as Navbar} from './Navbar/Navbar';
-export {default as Overlay} from './Overlay/Overlay';
-export {default as Drawer} from './Drawer/Drawer';
