+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Navbar } from "./Navbar/Navbar";
+export { default as Overlay } from "./Overlay/Overlay";
+export { default as Drawer } from "./Drawer/Drawer";
