--- src/data/portfolioData.js
@@ -1,105 +1,0 @@
-const data = {
-  items: [
-    {
-      _id: "jomalone",
-      title: "Jo Malone",
-      thumbnailUrl: "/images/jomalone.jpg",
-      description:
-        "At Perch Interactive, we were tasked by Jo Malone to create an immersive experience that merged digital and physical interactions. We created a display case that detected when a customer picked up a fragrance and delivered beautiful floral animations.",
-    },
-    {
-      _id: "leif",
-      title: "Leif",
-      thumbnailUrl: "/images/leif-macbook.png",
-      description:
-        "At Leif, I led a team of software engineers building technology to increase access to education. Leif's income share agreements have helped tens of thousands of students get an education with less financial risk.",
-    },
-    {
-      _id: "ibmresearch",
-      title: "IBM Research: Innovation Lab",
-      thumbnailUrl: "/images/thinklab.jpg",
-      description:
-        "At the IBM Research THINKLab, I created immersive experiences to demonstrate our research capabilities and to identify new opportunities with Fortune 500 partners. I was in charge of the life sciences, healthcare, and material sciences engagements that helped accelerate millions in new business.",
-    },
-    {
-      _id: "mobilebanking",
-      title: "Waypost: Mobile Banking",
-      thumbnailUrl: "/images/waypost-mobile.png",
-      description:
-        "Developed a mobile banking platform for the independent workforce. We focused on making it easy to set aside money for taxes, track expenses, and remind clients about unpaid invoices. Built on React Native with a Node.js backend and Plaid integration.",
-    },
-    {
-      _id: "voter",
-      title: "Vot-ER",
-      thumbnailUrl: "/images/voter-mockup.jpg",
-      description:
-        "Designed and created a mobile-responsive webapp for Vot-ER to help healthcare providers to register their patients to vote.",
-    },
-    {
-      _id: "foster",
-      title: "NYC Child Services: Foster Care",
-      thumbnailUrl: "./images/agingout.png",
-      thumbnailPosition: "left top",
-      description: `Worked with the NYC Administration for Child Services and to create a resources for foster children aging out of the system. Interviewed users who were formerly in the foster care system and created digital resources around job search, scholarships, and housing assistance.`,
-    },
-    {
-      _id: "waypost",
-      title: "Waypost: Web App",
-      thumbnailUrl: "/images/waypost-web.png",
-      thumbnailPosition: "left center",
-      description:
-        "In the Waypost web app, we reimagined how to make invoicing more intuitive and effective for the freelancers struggling to get paid on time. We redesigned invoicing work like an inbox, decreasing the number of clicks required to request payment and intelligently identifying which clients have overdue bills.",
-    },
-    {
-      _id: "ibmresearchcrm",
-      title: "IBM Research: CRM",
-      thumbnailUrl: "./images/compass.png",
-      thumbnailPosition: "left top",
-      description: `The Compass was originally developed as a standalone catalog of IBM Research projects. Over time, it expanded to become the central platform for managing client visits to the flagship innovation lab. Over two years it processed over 400 clients.\n\nI started this project as a way to distribute information about IBM Research projects around the world. As we continued operations in our innovation lab, we began to receive a much higher volume of visit requests and needed to process them systematically.\n\nAs tool lead, I regularly worked with our team and other internal stakeholders to map out requirements, develop features, and deploy them for use.`,
-    },
-    {
-      _id: "sentiment",
-      title: "Twitter Sentiment Analysis",
-      thumbnailUrl: "./images/sentiment.png",
-      thumbnailPosition: "right center",
-      description: `Created a small personal project that used sentiment analysis and probability theory to discover positive and negative influencers on Twitter.`,
-    },
-    {
-      _id: "sarmapper",
-      title: "Search and Rescue: Mapping",
-      thumbnailUrl: "./images/sarmapper.png",
-      thumbnailPosition: "left center",
-      description:
-        "Helps rescuers visualize probability models of lost person behavior. Used by dozens of search and rescue teams for briefings and planning operations, including state emergency management agencies.",
-    },
-    {
-      _id: "neutrogena",
-      title: "Neutrogena",
-      thumbnailUrl: "./images/neutrogena.jpg",
-      description: `Create an interactive retail display for Neutrogena that detected which products were picked up and delivered relevant information.`,
-    },
-    {
-      _id: "news-nlp",
-      title: "News Analytics",
-      thumbnailUrl: "./images/news/timeline.png",
-      thumbnailPosition: "left top",
-      description: `Created for an IBM challenge to integate Watson services. The application pulled scraped news articles and analyzed the natural language to visualize information across timeline, map, and entity relationships. `,
-    },
-    {
-      _id: "geovis",
-      title: "Location History Visualization",
-      thumbnailUrl: "./images/gps.png",
-      thumbnailPosition: "left center",
-      description: `Personal project that compiled and visualized my Google Maps location history across half a decade.`,
-    },
-    {
-      _id: "harrypotter",
-      title: "Harry Potter Fan Website",
-      thumbnailUrl: "./images/harrypotter.png",
-      thumbnailPosition: "top center",
-      description: `Created a Harry Potter fan website when I was 10 years old. Featured realistic simulations of several literary scenes.`,
-    },
-  ],
-};
-
-export default data;
