+++ src/pages/Home/components/Portfolio/Portfolio.tsx
@@ -1,0 +1,25 @@
+import PortfolioItem from "./PortfolioItem";
+import "./portfolio.css";
+import { motion } from "framer-motion";
+import type { PortfolioItem as PortfolioItemType } from "../../../../data/portfolioData";
+
+type PortfolioProps = {
+  items: PortfolioItemType[];
+  openProjectById: (projectId: string) => void;
+};
+
+function Portfolio({ items, openProjectById }: PortfolioProps) {
+  return (
+    <motion.div className="portfolio flex-container">
+      {items.map((item) => (
+        <PortfolioItem
+          openProjectById={openProjectById}
+          key={item._id}
+          item={item}
+        />
+      ))}
+    </motion.div>
+  );
+}
+
+export default Portfolio;
