+++ tailwind.config.ts
@@ -1,4 +1,4 @@
-/** @type {import('tailwindcss').Config} */
+import type { Config } from "tailwindcss";
 
 export default {
   content: ["./src/index.html", "./src/**/*.{js,ts,jsx,tsx}", "./App.css"],
@@ -6,4 +6,4 @@ export default {
     extend: {},
   },
   plugins: [],
-};
+} satisfies Config;
